refactor reviewManager and helpers
