+++ src/lib/SRSCardReview.ts
@@ -2,9 +2,6 @@ import type { ISRSCardReview, SRSCardReviewRow } from '#/lib/types';
 import type { ReviewLog } from 'ts-fsrs';
 import type { State } from 'ts-fsrs';
 
-/**
- *
- */
 export default class SRSCardReview implements ISRSCardReview {
   id: string;
   card_id: string;
